server/requirements.txt
# The API Server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart
python-dotenv

# AI & Agents
google-generativeai>=0.4.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
groq>=0.4.0

# Database
pymongo>=4.6.0
pinecone-client>=3.1.0

# Tools & Utilities
requests
pydantic>=2.0.0
Pillow
pytesseract
beautifulsoup4